Disallowid of 0 in deleteMaintainersForVersion(). Add a unit test to ensure that deleteMaintainersForVersion() behaves properly.

+++ include/maintainer.php
@@ -178,6 +178,9 @@ class maintainer
 
     function deleteMaintainersForVersion($oVersion)
     {
+        if(!$oVersion->iVersionId)
+            return FALSE;
+
         $hResult = query_parameters("DELETE from appMaintainers WHERE versionId='?'",
                                     $oVersion->iVersionId);
     }
